Refactor microphone button and speech page to use AppState for busy state management
