+++ ui/lib/components/audio/microphone_button.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 // Project imports:
+import 'package:ui/state/app_state.dart';
 import 'package:ui/state/speech_state.dart';
 
 class MicrophoneButton extends StatelessWidget {
@@ -19,24 +20,25 @@ class MicrophoneButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final appState = Provider.of<AppState>(context);
     final speechState = Provider.of<SpeechState>(context);
 
     return GestureDetector(
-      onLongPressStart: speechState.isBusy ? null : (_) => onStartRecording(),
-      onLongPressEnd: speechState.isBusy ? null : (_) => onStopRecording(),
+      onLongPressStart: appState.isBusy ? null : (_) => onStartRecording(),
+      onLongPressEnd: appState.isBusy ? null : (_) => onStopRecording(),
       child: Container(
         padding: const EdgeInsets.all(32),
         decoration: BoxDecoration(
           shape: BoxShape.circle,
           color: speechState.isRecording
               ? Theme.of(context).colorScheme.primary
-              : speechState.isBusy
+              : appState.isBusy
                   ? Theme.of(context).colorScheme.onSurface.withOpacity(0.5)
                   : Theme.of(context).colorScheme.secondary,
         ),
         child: Icon(
           Icons.mic,
-          color: speechState.isBusy
+          color: appState.isBusy
               ? Theme.of(context).colorScheme.onSurface.withOpacity(0.5)
               : Theme.of(context).colorScheme.onSecondary,
           size: 100,
