+++ ui/lib/state/speech_state.dart
@@ -4,11 +4,9 @@ import 'package:flutter/material.dart';
 class SpeechState extends ChangeNotifier {
   bool _microphonePermissionGranted = false;
   bool _isRecording = false;
-  bool _isBusy = false;
 
   bool get microphonePermissionGranted => _microphonePermissionGranted;
   bool get isRecording => _isRecording;
-  bool get isBusy => _isBusy;
 
   void setMicrophonePermissionGranted(bool granted) {
     _microphonePermissionGranted = granted;
@@ -17,11 +15,6 @@ class SpeechState extends ChangeNotifier {
 
   void setIsRecording(bool recording) {
     _isRecording = recording;
-    notifyListeners();
-  }
-
-  void setIsBusy(bool busy) {
-    _isBusy = busy;
     notifyListeners();
   }
 }
