+++ ui/test/components/audio/microphone_button_test.dart
@@ -7,19 +7,25 @@ import 'package:provider/provider.dart';
 
 // Project imports:
 import 'package:ui/components/audio/microphone_button.dart';
+import 'package:ui/state/app_state.dart';
 import 'package:ui/state/speech_state.dart';
 
 void main() {
+  late AppState appState;
   late SpeechState speechState;
 
   setUp(() {
+    appState = AppState();
     speechState = SpeechState();
   });
 
   Widget createMicrophoneButton(
       VoidCallback onStartRecording, VoidCallback onStopRecording) {
-    return ChangeNotifierProvider<SpeechState>.value(
-      value: speechState,
+    return MultiProvider(
+      providers: [
+        ChangeNotifierProvider(create: (_) => appState),
+        ChangeNotifierProvider(create: (_) => speechState),
+      ],
       child: MaterialApp(
         home: Scaffold(
           body: MicrophoneButton(
@@ -84,7 +90,7 @@ void main() {
       stopRecordingCalled = true;
     }
 
-    speechState.setIsBusy(true);
+    appState.setIsBusy(true);
     await tester
         .pumpWidget(createMicrophoneButton(onStartRecording, onStopRecording));
