+++ ui/test/state/speech_state_test.dart
@@ -15,7 +15,6 @@ void main() {
     test('initial values are correct', () {
       expect(speechState.microphonePermissionGranted, false);
       expect(speechState.isRecording, false);
-      expect(speechState.isBusy, false);
     });
 
     test('setMicrophonePermissionGranted', () {
@@ -26,11 +25,6 @@ void main() {
     test('setIsRecording', () {
       speechState.setIsRecording(true);
       expect(speechState.isRecording, true);
-    });
-
-    test('setIsBusy', () {
-      speechState.setIsBusy(true);
-      expect(speechState.isBusy, true);
     });
   });
 }
